+++ protobuf/uuidpb/map_test.go
@@ -1,10 +1,8 @@
 package uuidpb_test
 
 import (
-	"fmt"
 	"testing"
 
-	"github.com/dogmatiq/dapper"
 	. "github.com/dogmatiq/enginekit/protobuf/uuidpb"
 )
 
@@ -58,46 +56,12 @@ func TestMap(t *testing.T) {
 	}
 
 	expect = "bar"
-	for k, v := range m {
-		id := k.AsUUID()
-
+	for id, v := range m.All() {
 		if !id.Equal(k2) {
 			t.Errorf("unexpected key: got %q, want %q", id, k2)
 		}
 		if v != expect {
 			t.Errorf("unexpected value: got %q, want %q", v, expect)
 		}
-	}
-}
-
-func TestMapKey_DapperString(t *testing.T) {
-	t.Parallel()
-
-	subject := (&UUID{
-		Upper: 0xa967a8b93f9c4918,
-		Lower: 0x9a4119577be5fec5,
-	}).AsMapKey()
-
-	actual := dapper.Format(subject)
-	expect := "github.com/dogmatiq/enginekit/protobuf/uuidpb.MapKey [a967a8b9-3f9c-4918-9a41-19577be5fec5]"
-
-	if actual != expect {
-		t.Fatalf("got %q, want %q", actual, expect)
-	}
-}
-
-func TestMapKey_Format(t *testing.T) {
-	t.Parallel()
-
-	subject := (&UUID{
-		Upper: 0xa967a8b93f9c4918,
-		Lower: 0x9a4119577be5fec5,
-	}).AsMapKey()
-
-	expect := subject.String()
-	actual := fmt.Sprintf("%s", subject)
-
-	if actual != expect {
-		t.Fatalf("got %q, want %q", actual, expect)
 	}
 }